+++ audionimbus/CHANGELOG.md
@@ -81,7 +81,7 @@
 - Implement `Copy`, `Clone` for `SimulationError`.
 - Implement `Copy` for `ParameterValidationError`.
 - Implement `Clone` for `Spatialization`, `PathEffectSettings`, `VirtualSurroundEffectSettings`, `VirtualSurroundEffectParams`, `AmbisonicsBinauralEffectSettings`, `AmbisonicsBinauralEffectParams`, `AmbisonicsDecodeEffectSettings`, `AmbisonicsDecodeEffectParams` and `ReconstructorInputs`.
-- Implement `PartialEq`, `Eq` and `Hash` for `Scene`.
+- Implement `PartialEq`, `Eq` and `Hash` for `Scene` and `Simulator`.
 
 ### Removed